refactor: 사용하지는 않는 훅과 디렉토리 변경 반영
