+++ src/app/(search)/search/page.tsx
@@ -2,15 +2,14 @@
 
 import { mapAtom, markerAtom } from "@/src/app/_atom/atom";
 import { useRouter } from "next/navigation";
-import { useRecoilState, useSetRecoilState } from "recoil";
+import { useRecoilValue, useSetRecoilState } from "recoil";
 import InfoCard from "../_component/InfoCard";
 import styles from "../_component/InfoCard.module.css";
 import { IMarkers } from "../../_interfaces/interface";
 
 function Home() {
   const router = useRouter();
-  const [markerRecoil, setMarkerRecoil] =
-    useRecoilState<IMarkers[]>(markerAtom);
+  const markerRecoil = useRecoilValue<IMarkers[]>(markerAtom);
   const mapRecoil = useSetRecoilState<IMarkers>(mapAtom);
   const onDivClicked = (data: IMarkers) => {
     mapRecoil(data);
